+++ CHANGELOG.md
@@ -27,6 +27,12 @@
 
 ## Unreleased
 
+### Added
+
+- In `KonnectGatewayControlPlane` fields `Status.Endpoints.ControlPlaneEndpoint`
+  and `Status.Endpoints.TelemetryEndpoint` are filled with respective values from Konnect.
+  [#1415](https://github.com/Kong/gateway-operator/pull/1415)
+
 ### Changes
 
 - Added `spec.watchNamespace` field to `ControlPlane` CRD to allow watching resources